refactor: extract details-state-mixin from base
